fixed typo issues and moved data for announcements into constants folder. Also renamed Home page folder to Announcements
