+++ frontend/VolunteerOne/constants/announcements.js
@@ -1,0 +1,34 @@
+export const dataList = [
+  {
+    id: 0,
+    organization: "American Red Cross",
+    name: "New Event",
+    subject: "Help pack disaster relief bags",
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+    date: "",
+    location: "",
+    timePosted: new Date("2023-03-01T00:00:00"),
+    type: "event",
+  },
+  {
+    id: 1,
+    organization: "Habitat for Humanity",
+    name: "New Event",
+    subject: "Build and improve homes",
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+    date: "",
+    location: "",
+    timePosted: new Date("2023-02-02T00:00:00"),
+    type: "event",
+  },
+  {
+    id: 2,
+    organization: "Global Volunteers",
+    announcement:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et enim ad minim veniam, quis nostrud exercitation ullamco laboris ",
+    type: "announcement",
+    timePosted: new Date("2021-03-01T00:00:00"),
+  },
+];
